Refactor WorkflowExecution and related types to be globally available
Just like we're doing with the events/
