+++ src/lib/services/terminate-service.ts
@@ -1,8 +1,6 @@
 import { requestFromAPI } from '$lib/utilities/request-from-api';
 import { routeForApi } from '$lib/utilities/route-for-api';
 
-import type { WorkflowExecution } from '$lib/models/workflow-execution';
-
 type TerminateWorkflowOptions = {
   workflow: WorkflowExecution;
   namespace: string;
