+++ src/lib/utilities/get-visible-events.ts
@@ -1,5 +1,3 @@
-import type { WorkflowExecution } from '$lib/models/workflow-execution';
-
 import { eventTypeInCategory } from './get-event-categorization';
 
 export const getVisibleEvents = async (
