+++ src/workflow.d.ts
@@ -1,0 +1,51 @@
+type WorkflowExecutionStatus = import('$types').WorkflowExecutionStatus;
+type ListWorkflowExecutionsResponse =
+  import('$types').ListWorkflowExecutionsResponse;
+
+type WorkflowExecutionAPIResponse = Optional<
+  DescribeWorkflowExecutionResponse,
+  'executionConfig' | 'pendingActivities' | 'pendingChildren'
+>;
+
+type WorkflowStatus =
+  | 'Running'
+  | 'TimedOut'
+  | 'Completed'
+  | 'Failed'
+  | 'Completed'
+  | 'ContinuedAsNew'
+  | 'Canceled'
+  | 'Terminated'
+  | null;
+
+type WorkflowType = string | null;
+
+type WorkflowExecutionFilters = {
+  type: WorkflowType;
+  status: WorkflowStatus;
+};
+
+type FilterParameters = {
+  workflowId?: string;
+  workflowType?: string;
+  executionStatus?: WorkflowStatus;
+  timeRange?: Duration | string;
+  query?: string;
+};
+
+type ArchiveFilterParameters = Omit<FilterParameters, 'timeRange'> & {
+  closeTime: Duration | string;
+};
+
+type WorkflowExecution = {
+  name: string;
+  id: string;
+  runId: string;
+  startTime: string;
+  endTime: string;
+  status: WorkflowExecutionStatus;
+  taskQueue?: string;
+  historyEvents: Long;
+  pendingActivities: PendingActivity[];
+  url: string;
+};
